requirements.txt
Django==2.2.1
django-ckeditor==5.7.1
django-cleanup==3.2.0
django-crispy-forms==1.7.2
django-js-asset==1.2.2